Change the association in GraphQL to match the ERD
for request types
